Curry.jl

Documentation for Curry.jl

Curry.Bind — Type

Represent a function call, with partially bound arguments.

b = Bind(+, (1, 2))
b()

# output

3

b = Bind(*, ("hello", nothing))
b(", world")

# output

"hello, world"

Bind(f, (g(), h())) is like :(f(g(), h())) but f g and h are lexically scoped, and g() and h() are evaluated eagerly.

Curry.interleave — Method

Return a Tuple that interleaves args into the nothing slots of slots.

Curry.interleave((:a, nothing, :c, nothing), (12, 34))

# output

(:a, 12, :c, 34)

Use Some to escape nothing

Curry.interleave((:a, Some(nothing), :c, nothing), (34,))

# output

(:a, nothing, :c, 34)