Curry.jl

Documentation for Curry.jl

Curry.fix — Method

`fix(f, a, b)`
`fix(f, args...; kw...)`

The fix function partially evaluates f by fix some of its arguments. Positional arguments of f that should not be bound are indicated by passing nothing to fix at the respective position.

julia> using Curry: fix

julia> b = fix(+, 1, 2); # no nothing, all arguments bound

julia> b()
3

julia> b = fix(*, "hello", nothing); # only first argument bound

julia> b(", world")
"hello, world"

julia> b = fix(=>, nothing, 1); # second argument bound

julia> b("one")
"one" => 1

julia> b = fix(isapprox, nothing, nothing, atol=100); # only atol keyword bound

julia> b(10, 20)
true

julia> b(10, 20, atol=1) # keywords can be reassigned on the fly
false

Curry.interleave — Method

Return a Tuple that interleaves args into the nothing slots of slots.

Curry.interleave((:a, nothing, :c, nothing), (12, 34))

# output

(:a, 12, :c, 34)

Use Some to escape nothing

Curry.interleave((:a, Some(nothing), :c, nothing), (34,))

# output

(:a, nothing, :c, 34)

Curry.Fix — Type

Represent a function call, with partially bound arguments.