FixArgs.jl

FixArgs.jl

Introduction

Examples of Base.Fix2

Symbolic computation and lazy evaluation

Base.Iterators

Base.Rational

Fixed-Point Numbers and "static" arguments

Pure-imaginary type and Base.Complex

Introduction

This package began as an exploration in generalizing Base.Fix1 and Base.Fix2. These types are ways to represent a particular forms of anonymous functions. Let's illustrate. We'll use the string function in Base, which concatenates the string representations of its arguments:

julia> string("first ", "second")
"first second"

Now, to construct and use the Fix1 and Fix2 types:

julia> using Base: Fix1, Fix2

julia> f1 = Fix1(string, "one then ")
(::Base.Fix1{typeof(string),String}) (generic function with 1 method)

julia> f1("two")
"one then two"

The function-call behavior of Fix1(f, bind) is the same as x -> f(bind, x).

Similarly,

julia> f2 = Fix2(string, " before two")
(::Base.Fix2{typeof(string),String}) (generic function with 1 method)

julia> f2("one")
"one before two"

The function-call behavior of Fix2(f, bind) is the same as x -> f(x, bind).

The key point of the Fix1 and Fix2 types is that methods can dispatch on

the type of f

the type of bind

the position of bind within the function call

Dispatch is not tenable with anonymous functions. Let's illustrate while moving to a more practical example using == instead of string.

julia> f1 = x -> x == 0
#1 (generic function with 1 method)

julia> f2 = Fix1(==, 0)
(::Base.Fix1{typeof(==),Int64}) (generic function with 1 method)

Now define the "same" things again:

julia> f3 = x -> x == 0
#3 (generic function with 1 method)

julia> f4 = Fix1(==, 0)
(::Base.Fix1{typeof(==),Int64}) (generic function with 1 method)

The types of both the Fix1 values is the same:

julia> typeof(f2) === typeof(f4)
true

But each anonymous function definition introduces a new type with an opaque name:

julia> typeof(f1), typeof(f3)
(Main.ex-BaseFix.var"#1#2", Main.ex-BaseFix.var"#3#4")

A new anonymous function is always given a unique type, which allows methods to specialize on the specific anonymous function passed as an argument, but does not "permit" dispatch. To be more accurate, as far as dispatch is concerned, the type of anonymous functions is not special:

julia> foo(::typeof(f1)) = "f1"
foo (generic function with 1 method)

julia> foo(::typeof(f3)) = "f3"
foo (generic function with 2 methods)

julia> foo(f1)
"f1"

julia> foo(f3)
"f3"

But really we'd like to use a type that is less opaque and furthermore is "structural" in some ways, rather than purely "nominal".

Examples of Base.Fix2

Where is it useful to dispatch on these special functions? Because Base does not export and does not document these types, there aren't methods in the Julia ecosystem.

But these types are constructed with, for example, ==(3) or in([1, 2, 3]). A type like these is useful as a predicate to pass to higher-order functions, e.g. findfirst(==(3), some_array) to find the first element that equals 3. Brevity asside, these types are useful to define more efficient methods of generic higher-order functions. For example, take a specific method of the findfirst function:

findfirst(p::Union{Fix2{typeof(isequal),Int},Fix2{typeof(==),Int}}, r::OneTo{Int}) =
    1 <= p.x <= r.stop ? p.x : nothing

The fallback for findfirst (triggered by e.g. findfirst(x->x==3, 1:10) instead of findfirst(==(3), 1:10)) would produce the same (correct) answer, but the method above will be quicker.

Dispatching on the structure of the predicate function enables a certain form of symbolic computation.

Symbolic computation and lazy evaluation

This package provides a generalization of Fix1 and Fix2 in a few ways:

A function of any positional arity can be used, and any number of its arguments can be bound, allowing the remaining arguments to be provided later.

A function can have its keyword arguments bound.

The function x -> f(x, b) is represented with types:

a Lambda to represent function (args -> body)

a Call to represent the function call (f(...)) in the body

a ArgPos to represent the x in the body of the lambda function

The third generalization is powerful, because it's effectively the lambda calculus.

It is worth considering first just Call, which can serve the purpose of representing a delayed function call evaluation. If you prefer, you may also consider a thunk () -> foo(1, 2), which would be a Lambda (with no arguments) and a Call that does not mention any "free variables".

If laziness is all that is needed, then defining a Julia anonymous function will do the job. But this package allows an additional benefit since methods can dispatch on details of the lazy call.

In many domains, new types are introduced to represent this pattern.

Base.Iterators

Base.Generator consists of two fields f and iter. This can be taken as a representation of map(f, iter):

julia> using FixArgs

julia> gen = let f = string, iter = 1:10
           @xquote map(f, iter)
       end
Call(Some(map), FrankenTuple((Some(string), Some(1:10)), NamedTuple()))

It's certainly less nice to look at than Base.Generator{UnitRange{Int64}, typeof(string)}(string, 1:10). Better UX / ergonomics are be possible by defining a type alias:

const MyGenerator{F, I} = FixArgs.Call{Some{typeof(map)}, FixArgs.FrankenTuples.FrankenTuple{Tuple{Some{F}, Some{I}, (), Tuple{}}}

That is quite unsightly, and there are quite a few internals leaking out. We can use a macro instead:

julia> const MyGenerator{F, I} = @xquoteT map(::F, ::I)
FixArgs.Call{Some{typeof(map)},FrankenTuples.FrankenTuple{Tuple{Some{F},Some{I}},(),Tuple{}}} where I where F

It should be made convenient to defining constructors and show methods that correspond with the type alias.

To evaluate the call (i.e. "collect the iterator"):

julia> xeval(gen)
10-element Array{String,1}:
 "1"
 "2"
 "3"
 "4"
 "5"
 "6"
 "7"
 "8"
 "9"
 "10"

This example is actually circular. The evaluation of the map call is done in terms of Generator! The definition:

map(f, A) = collect(Generator(f,A))

Breaking this circularity is possible by defining

function iterate(gen::(@xquoteT map(::F, ::I))) where F, I
    f = FixArgs.xeval(gen.args[1])  # not the prettiest thing right now...
    iter = FixArgs.xeval(gen.args[2])
    # ...
end

and might also require a separation of the purposes of collect and map. See this issue.

Many types in Base.Iterators can be seen as lazy calls of existing functions. Base.Iterators.Filter(flt, itr) could be replaced with @xquote filter(flt, itr). And the dispatches done on these types to enable the existing symbolic computation, e.g.:

reverse(f::Filter) = Filter(f.flt, reverse(f.itr))

Base.Iterators.Flatten, which defines a convenience function

flatten(itr) = Flatten(itr)

could be written in terms of a function flatten with no methods. However, it is perhaps better seen as @xquote reduce(vcat, it)

Base.Rational

What is Rational but lazy division on integers?

julia> 1/9 * 3/2 # eager division
0.16666666666666666

julia> using FixArgs

julia> (@xquote 1/9) * (@xquote 3/2)
ERROR: MethodError: no method matching *(::FixArgs.Call{Some{typeof(/)},FrankenTuples.FrankenTuple{Tuple{Some{Int64},Some{Int64}},(),Tuple{}}}, ::FixArgs.Call{Some{typeof(/)},FrankenTuples.FrankenTuple{Tuple{Some{Int64},Some{Int64}},(),Tuple{}}})
Closest candidates are:
  *(::Any, ::Any, !Matched::Any, !Matched::Any...) at operators.jl:538

Of course, we have to do some more work.

using Base: divgcd

function Base.:*(
        x::(@xquoteT ::T / ::T),
        y::(@xquoteT ::T / ::T),
        ) where {T}
    xn, yd = divgcd(something(x.args[1]), something(y.args[2]))
    xd, yn = divgcd(something(x.args[2]), something(y.args[1]))
    ret = @xquote $(xn * yn) / $(xd * yd) # TODO use `unsafe_rational` and `checked_mul`
    ret
end

Now, try again:

julia> q = (@xquote 1/9) * (@xquote 3/2)
Call(Some(/), FrankenTuple((Some(1), Some(6)), NamedTuple()))

julia> map(xeval, q.args) # make numerator and denominator plainly visible
ERROR: MethodError: no method matching xeval(::Some{Int64})
Closest candidates are:
  xeval(::Some, !Matched::FixArgs.Context) at /home/runner/work/FixArgs.jl/FixArgs.jl/src/eval.jl:17
  xeval(::Some, !Matched::Nothing) at /home/runner/work/FixArgs.jl/FixArgs.jl/src/eval.jl:18
  xeval(!Matched::FixArgs.ArgPos{i}, !Matched::FixArgs.Context{Nothing,P}) where {i, P} at /home/runner/work/FixArgs.jl/FixArgs.jl/src/eval.jl:13
  ...

compare with using // to construct a Base.Rational:

julia> 1//9 * 3//2
1//6

Finally, because we have encoded the relationship between this "new" rational type, and /, we can do:

julia> xeval(q)
0.16666666666666666

We could define an alias:

const MyRational{T} = @xquoteT(::T / ::T)

FixArgs.Call{Some{typeof(/)},FrankenTuples.FrankenTuple{Tuple{Some{T},Some{T}},(),Tuple{}}} where T

which would also enforce the same type for both the numerator and denominator, as is the case of Base.Rational.

julia> sizeof(MyRational{Int32})
8

Occasionally, a user might find this to be a limitation, yet they would still like to use some of the generic algorithms that might apply.

The fields of Base.Rational are num and den. They have to be named since that's all that gives the fields any meaning at all. In our type, however, instead of naming the fields they can be distinguished by the role they play with respect to the / function.

Fixed-Point Numbers and "static" arguments

A fixed-point number is just a rational number with a specified denominator. If we have a large array of fixed-point numbers with the same denominator, we certainly do not want to store the denominator repeatedly.

And we want to ensure constant propagation happens, too.

So we can "bake in" some values (Base.isbitstype) into the type of Call itself!

In other words, what is a fixed-point number but lazy division with a static denominator? Here is an example that models Fixed{Int8,7} from FixedPointNumbers.jl. The macros use the notation V::::S to mark an argument V as "static". Also note the use of $ to escape subexpressions.

julia> using FixArgs

julia> MyFixed{Int8, 128} === typeof(MyQ0f7(3))
true

julia> function Base.:+(a::MyFixed{N,D}, b::MyFixed{N,D})::MyFixed{N,D} where {N, D}
           n = something(a.args[1]) + something(b.args[1])
           return (@xquote $(N(n)) / D::::S)
       end

julia> xeval(MyQ0f7(3) + MyQ0f7(2)) === 5/128
true

julia> sizeof(MyFixed)
ERROR: Argument is an incomplete Call type and does not have a definite size.

julia> sizeof(Int8)
1

And the generated code appears to be equivalent between

using FixedPointNumbers
look_inside_1(x, y) = reinterpret(Fixed{Int8, 7}, Int8(x)) + reinterpret(Fixed{Int8, 7}, Int8(y))

and

look_inside_2(x, y) = MyQ0f7(x) + MyQ0f7(y)

Pure-imaginary type and Base.Complex

Now that we can make some arguments static, we can introduce a meaningful example where the lazy call might not be valid to begin with. You can define a type such that xeval raises MethodError and still represent the computation symbolically. The Julia ecosystem goes to great lengths to find the right generic functions and to ensure that all methods defined on generic functions are semantically compatible. This effort enables generic programming and interoperability. You can define a type A in terms of a function f and a type B even if it may not make sense to define a new method of f on B.

Here is an over-the-top example:

julia> using FixArgs

julia> struct ImaginaryUnit end    # if we want to be really cute, can do `@xquote sqrt((-1)::::S)`

julia> const Imaginary{T} = @xquoteT ::T * ::ImaginaryUnit
FixArgs.Call{Some{typeof(*)},FrankenTuples.FrankenTuple{Tuple{Some{T},Some{Main.ex-Imaginary.ImaginaryUnit}},(),Tuple{}}} where T

julia> Imaginary(x) = @xquote x * $(ImaginaryUnit())   # note escaping
FixArgs.Call{Some{typeof(*)},FrankenTuples.FrankenTuple{Tuple{Some{T},Some{Main.ex-Imaginary.ImaginaryUnit}},(),Tuple{}}} where T

note that if we assume we have no Base.Complex or anything like it, we don't have a way to further evaluate:

julia> xeval(Imaginary(3))
ERROR: MethodError: no method matching *(::Int64, ::Main.ex-Imaginary.ImaginaryUnit)
Closest candidates are:
  *(::Any, ::Any, !Matched::Any, !Matched::Any...) at operators.jl:538
  *(::T, !Matched::T) where T<:Union{Int128, Int16, Int32, Int64, Int8, UInt128, UInt16, UInt32, UInt64, UInt8} at int.jl:87
  *(::Union{Int16, Int32, Int64, Int8}, !Matched::BigInt) at gmp.jl:538
  ...

We represented pure imaginary numbers as lazy multiplication of numbers and a singleton type ImaginaryUnit, and it is basically as if we had defined

struct Imaginary{T}
    _::T
end

Let's just go ahead and represent complex numbers too:

julia> # const MyComplex{R, I} = @xquoteT ::R + (::I * ::ImaginaryUnit) # TODO this macro doesn't work
       MyComplex(r, i) = @xquote r + i * $(ImaginaryUnit())
MyComplex (generic function with 1 method)

Note this monster of a type has the same size as Base.Complex:

julia> sizeof(Complex(1, 2))
16

julia> sizeof(MyComplex(1, 2))
16

and layout too:

julia> reinterpret(Int64, [Complex(1, 2)])
2-element reinterpret(Int64, ::Array{Complex{Int64},1}):
 1
 2

julia> reinterpret(Int64, [MyComplex(1, 2)])
2-element reinterpret(Int64, ::Array{FixArgs.Call{Some{typeof(+)},FrankenTuples.FrankenTuple{Tuple{Some{Int64},FixArgs.Call{Some{typeof(*)},FrankenTuples.FrankenTuple{Tuple{Some{Int64},Some{Main.ex-Imaginary.ImaginaryUnit}},(),Tuple{}}}},(),Tuple{}}},1}):
 1
 2

Of course, there are many different types that would all be mathematically equivalent by swapping the arguments to + or *. Note that swapping the arguments to + would give a different memory layout.

FixArgs.ArgPos — Type

Within the body of a Lambda, represent a formal positional parameter of thatLambda.

FixArgs.Arity — Type

Represent the arity of a Lambda.

Currently, only represents a fixed number of positional arguments, but may be generalized to include optional and keyword arguments.

P is 0, 1, 2, ... KW is always NoKeywordArguments, and may be extended in the future.

FixArgs.Call — Type

A call "f(args...)". args may represent both positional and keyword arguments.

FixArgs.Context — Type

terms are evaluated with respect to a Context A Context is an associations between bound variables and values, and they may be nested (via the parent field).

FixArgs.Lambda — Type

A lambda expression "args -> body"

FixArgs.ParentScope — Type

Nest ArgPos in ParentScopes to represent a reference to the formal parameters of a "parent" function. Forms a unary representation.

Related: [De Bruijn indices]https://en.wikipedia.org/wiki/DeBruijnindex

FixArgs.TypedExpr — Type

Roughly mirror Base.Expr, except that the the head of the expression (encoded in the head field) can be dispatched on.

This is only used in an intermediate representation of this package.

Note that Expr and TypedExpr are constructed slightly differently. Each argument of an Expr is an argument to Expr, whereas all arguments of a TypedExpr are passed as one argument (a tuple) to TypedExpr

e.g.

Expr(:call, +, 1, 2) corresponds to TypedExpr(Val{:call}(), (+, 1, 2))

FixArgs.isexpr — Method

isexpr(expr) -> Bool
isexpr(expr, head) -> Bool

Checks whether given value isa Base.Expr and if further given head, it also checks whether the head matches expr.head.

Examples

julia> using ExprParsers
julia> EP.isexpr(:(a = hi))
true
julia> EP.isexpr(12)
false
julia> EP.isexpr(:(f(a) = a), :(=))
true
julia> EP.isexpr(:(f(a) = a), :function)
false

FixArgs.lc_expr — Method

Convert a ::TypedExpr to a Lambda-Call expression

FixArgs.normalize_lambda_1_arg — Method

normalize :(x -> body) into :((x,) -> body)

FixArgs.postwalk — Method

postwalk(f, expr)

Applies f to each node in the given expression tree, returning the result. f sees expressions after they have been transformed by the walk.

See also: prewalk.

FixArgs.prewalk — Method

prewalk(f, expr)

Applies f to each node in the given expression tree, returning the result. f sees expressions before they have been transformed by the walk, and the walk will be applied to whatever f returns.

This makes prewalk somewhat prone to infinite loops; you probably want to try postwalk first.

FixArgs.relabel_args — Function

α-conversion in λ-calculus

labeler(x) produces a Symbol or similar from x.referent_depth x.antecedent_depth x.arg_i x.sym – name before relabeling

x.referent_depth - x.antecedent_depth is number of ->s that are between the evaluation site and the definition site

FixArgs.uneval — Method

Given a value, produce an expression that when eval'd produces the value.

e.g.

julia> eval(uneval(Expr(:my_call, :arg1, :arg2)))
:($(Expr(:my_call, :arg1, :arg2)))

julia> eval(eval(uneval(:(sqrt(9)))))
3.0

This function is used to return expressions from this package's macros. This is likely not a well-posed problem to begin with. Related issue.

Note the special case for :(esc(x)).

FixArgs.xapply — Function

apply a Lambda expression to arguments.

FixArgs.xeval — Method

evaluate a Lambda-Call expression

Currently only works on Call expressions.

FixArgs.@fix — Macro

A convenience macro that implements the syntax of this PR

@fix f(_, b) is the equivalent of x -> f(x, b)

FixArgs.@quote_some — Macro

This macro is used to debug and introspect the escaping behavior of @xquote

julia> dump(let x = 9
       @xquote sqrt(x)
       end)
Expr
    head: Symbol call
    args: Array{Any}((2,))
        1: sqrt (function of type typeof(sqrt))
        2: Int64 9

FixArgs.@xquote — Macro

Transform julia syntax into a Lambda-Call expression.

FixArgs.@xquoteT — Macro

The types produced by this package are unwieldly. This macro permits a convenient syntax, e.g. @xquoteT func(::Arg1Type, ::Arg2Type) to represent types.

let func = identity, arg = 1
    typeof(@xquote func(arg)) == @xquoteT func(::typeof(arg))
end

# output

true

If an argument is "static", then it is part of the type, and the value is annotated as illustrated:

@xquoteT string(123::::S)

# output

FixArgs.Call{Some{typeof(string)},FrankenTuples.FrankenTuple{Tuple{Val{123}},(),Tuple{}}}